requirements.txt
Click==7.0
Flask==1.0.2
Flask-Limiter==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.1
limits==1.3
MarkupSafe==1.1.1
Pillow==6.0.0
six==1.12.0
Werkzeug==0.15.2
webcolors==1.8.1
